+++ test/compute-engine/incomplete-expressions.test.ts
@@ -99,13 +99,13 @@ describe('basic', () => {
       `["Equal", "x", ["Error", "'missing'"]]`
     );
     expect(parse('2 \\times 2 = ')).toMatchInlineSnapshot(
-      `["Equal", ["Multiply", 2, 2], ["Error", "'missing'"]]`
+      `["Equal", 4, ["Error", "'missing'"]]`
     );
     expect(parse('x+1=')).toMatchInlineSnapshot(
       `["Equal", ["Add", "x", 1], ["Error", "'missing'"]]`
     );
     expect(parse('2+1\\le')).toMatchInlineSnapshot(
-      `["LessEqual", ["Add", 1, 2], ["Error", "'missing'"]]`
+      `["LessEqual", 3, ["Error", "'missing'"]]`
     );
   });
 });